# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Restaurant API",
    "version": "1.0.0",
    "description": "API for managing restaurant operations"
  },
  "servers": [
    {
      "url": "http://localhost:8888",
      "description": "Local server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/reviews": {
      "get": {
        "tags": ["Reviews"],
        "summary": "Reviews",
        "description": "The authenticated and authorized user can view reviews",
        "responses": {
          "200": {
            "description": "Reviews.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "userId": {
                            "type": "object",
                            "properties": {
                              "username": {
                                "type": "string"
                              },
                              "email": {
                                "type": "string"
                              },
                              "phone": {
                                "type": "string"
                              }
                            }
                          },
                          "menuItemId": {
                            "type": "object",
                            "properties": {
                              "name": {
                                "type": "string"
                              },
                              "price": {
                                "type": "array",
                                "items": {
                                  "type": "string"
                                }
                              }
                            }
                          },
                          "rating": {
                            "type": "number"
                          },
                          "comment": {
                            "type": "string"
                          },
                          "createdAt": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "You dont have permission."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      },
      "post": {
        "tags": ["Reviews"],
        "summary": "add Review",
        "description": "The authenticated user can add a review.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "menuItemId": {
                    "type": "string"
                  },
                  "rating": {
                    "type": "number"
                  },
                  "comment": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Review added successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "userId": {
                            "type": "object",
                            "properties": {
                              "username": {
                                "type": "string"
                              },
                              "email": {
                                "type": "string"
                              },
                              "phone": {
                                "type": "string"
                              }
                            }
                          },
                          "menuItemId": {
                            "type": "object",
                            "properties": {
                              "name": {
                                "type": "string"
                              },
                              "price": {
                                "type": "array",
                                "items": {
                                  "type": "string"
                                }
                              }
                            }
                          },
                          "rating": {
                            "type": "number"
                          },
                          "comment": {
                            "type": "string"
                          },
                          "createdAt": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/api/reviews/{menuItemId}": {
      "get": {
        "tags": ["Reviews"],
        "summary": "Reviews",
        "description": "The authenticated user can view the reviews of menu items",
        "parameters": [
          {
            "in": "path",
            "name": "menuItemId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "menu item id."
          }
        ],
        "responses": {
          "200": {
            "description": "Reviews.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "userId": {
                            "type": "object",
                            "properties": {
                              "username": {
                                "type": "string"
                              },
                              "email": {
                                "type": "string"
                              },
                              "phone": {
                                "type": "string"
                              }
                            }
                          },
                          "menuItemId": {
                            "type": "object",
                            "properties": {
                              "name": {
                                "type": "string"
                              },
                              "price": {
                                "type": "array",
                                "items": {
                                  "type": "string"
                                }
                              }
                            }
                          },
                          "rating": {
                            "type": "number"
                          },
                          "comment": {
                            "type": "string"
                          },
                          "createdAt": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/api/reviews/{id}": {
      "put": {
        "tags": ["Reviews"],
        "summary": "Update reviews",
        "description": "The authenticated user can update his reviews.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "review id."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "rating": {
                    "type": "number"
                  },
                  "comment": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Order updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "You are not authorized to update this review."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      },
      "delete": {
        "tags": ["Reviews"],
        "summary": "Delete Review",
        "description": "The authenticated user and authorized admin can delete the review.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "review id."
          }
        ],
        "responses": {
          "200": {
            "description": "Review deleted successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "You are not authorized to update this review."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    }
  }
}